print first 50 fibonacci numbers starting from 1 and 2

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,19 +9,19 @@
 int main(void)
 {
 	int i = 3;
-	int next;
-	long n1 = 1;
-	long n2 = 2;
+	long int next;
+	long int n1 = 1;
+	long int n2 = 2;
 
-	printf("%d, ", n1);
-	printf("%d, ", n2);
+	printf("%ld, ", n1);
+	printf("%ld, ", n2);
 
 	for (i = 3; i <= 50; i++)
 	{
 		next = n1 + n2;
 		n1 = n2;
 		n2 = next;
-		printf("%d, ", next);
+		printf("%ld, ", next);
 	}
 	printf("\n");
